requirements.txt
streamlit
pvporcupine
pyaudio
openai
pyttsx3
requests
beautifulsoup4
python-dotenv
pyyaml